+++ spec/mongoid/properties/host_spec.rb
@@ -13,7 +13,7 @@ describe Mongoid::Shell::Properties::Host do
     klass.new
   end
   it 'raises an exception when the session is not connected' do
-    if ::Mongoid::Compatibility::Version.mongoid5?
+    if ::Mongoid::Compatibility::Version.mongoid5? || ::Mongoid::Compatibility::Version.mongoid6?
       allow(Mongoid.default_client.cluster).to receive(:servers).and_return([])
     else
       allow(Mongoid.default_session.cluster).to receive(:nodes).and_return([])