+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ============
 
 * Your contribution here.
+* [#11](https://github.com/mongoid/mongoid-shell/pull/11) - Compatibility with Mongoid 6 - [@jbach](https://github.com/jbach).
 
 0.4.3 (6/1/2016)
 ================